Create 1
This code defines an array called notes to store the notes. The addNote() function is used to add a new note to the notes array, and the displayNotes() function is used to display all the notes in the array on the webpage. The deleteNote() function is used to delete a note from the notes array, and the setReminder() function is used to set a reminder for a note.